% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots.R
\name{bot_eq}
\alias{bot_eq}
\title{Bot that plays according to a specified equilibrium}
\usage{
bot_eq(game, player, eq = game$eq.li[[1]], eq.tables = gtree::eq_tables(game
  = game, eq.li = list(eq)), name = "eq_bot")
}
\arguments{
\item{game}{the game object}

\item{player}{the player number of this bot}

\item{eq}{an equilibrium, typically an element of game$eq.li}
}
\description{
Bot that plays according to a specified equilibrium
}
\seealso{
Other Bots: \code{\link{bot_random}},
  \code{\link{bot_tables}}, \code{\link{make_bots}},
  \code{\link{new_wp}}, \code{\link{play_bot_vs_bot}}
}
