% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots.R
\name{bot_mixture}
\alias{bot_mixture}
\title{Bot that mixes between different bots}
\usage{
bot_mixture(game, player, child_bots, prob = NULL, ...)
}
\arguments{
\item{game}{the game object}

\item{player}{the player number of this bot}

\item{child_bots}{A list of child pots}

\item{prob}{A vector of weights for each child bot. If \code{NULL} (default) all are equally likely.}
}
\description{
The first time the bot is called
for a particular player in a play
He picks a child bot randomly.
Then it continues with that child bot
for this player the whole play.
}
\details{
If you use \code{\link{make_bots}} or call
repeatedly \code{bot_mixture} to generate mixture bots
for each player, the child bots will be independently
drawn for each player.

Instead, if bot1 is a mixture bot for player 1
and you create a mixture bot for player 2 by
\code{bot2 = bot1\nbot2$player = 2}
then bot2 will select in every play the same
child bot than bot1.
}
\seealso{
Other Bots: \code{\link{bot_eq}}, \code{\link{bot_pop}},
  \code{\link{bot_random}}, \code{\link{bot_tables}},
  \code{\link{make_bots}}, \code{\link{play_bot_vs_bot}}
}
