% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_pop.R
\name{bot_pop}
\alias{bot_pop}
\title{Bot who mimics the average player population}
\usage{
bot_pop(game, player, pps, alt.bot = NULL, alt.bot.count = 5,
  name = "pop_bot", alt.bot.fun = bot_random, ...)
}
\arguments{
\item{game}{the game object}

\item{player}{the player number of this bot}

\item{pps}{a pps object, can be extended during play.}

\item{alt.bot}{a bot who will be aksed if there are too few observations in the popoulation}

\item{alt.bot.count}{we assume that we already have this many observations for alt.bot. This determines the probability to draw from the alt.bot instead of the population}
}
\description{
Draws actions from previous actions stored
in population play summary (pps) object
}
\seealso{
Other Bots: \code{\link{bot_eq}},
  \code{\link{bot_mixture}}, \code{\link{bot_random}},
  \code{\link{bot_tables}}, \code{\link{make_bots}},
  \code{\link{play_bot_vs_bot}}

Other population play functions: \code{\link{new_pps}},
  \code{\link{pps_add_play_actions}},
  \code{\link{pps_rearrange}}
}
