% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bots.R
\name{bot_tables}
\alias{bot_tables}
\title{Bot whose actions are determined by key-action tables}
\usage{
bot_tables(game, player, tables, name = "table_bot", ...)
}
\arguments{
\item{game}{the game object}

\item{player}{the player number of this bot}

\item{tables}{a list of tables for each action. The result of \code{eq_tables} is}
}
\description{
Bot whose actions are determined by key-action tables
}
\seealso{
Other Bots: \code{\link{bot_eq}}, \code{\link{bot_pop}},
  \code{\link{bot_random}}, \code{\link{make_bots}},
  \code{\link{new_wp}}, \code{\link{play_bot_vs_bot}}
}
