% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{deploy_webplay_example}
\alias{deploy_webplay_example}
\title{Deploys an example app to local directory}
\usage{
deploy_webplay_example(example = c("UltimatumGame", "KuhnPoker")[1],
  dest.dir = file.path(getwd(), example))
}
\arguments{
\item{example}{Name of the example. Current options are \n\t1. \code{"UltimatumGame"} a simple introductionary app and \n\t2.\code{"KuhnPoker"} a more complex app that implements a \code{\link{bot_pop}} to play against the population of earlier players.}

\item{dest.dir}{The destination directory}
}
\description{
Deploys an example app to local directory
}
