% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_app.R
\name{get_wp}
\alias{get_wp}
\title{Gets the web play object of the current app instance}
\usage{
get_wp(app = getApp())
}
\description{
Gets the web play object of the current app instance
}
\seealso{
set_wp_for_app

Other Web Play: \code{\link{new_wp}},
  \code{\link{set_wp_for_app}}, \code{\link{wpDevelApp}},
  \code{\link{wp_copy}}, \code{\link{wp_developer_ui}},
  \code{\link{wp_reset}}, \code{\link{wp_set_to_play}}
}
