% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_pop.R
\name{new_pps}
\alias{new_pps}
\title{Create a new empty population play summary}
\usage{
new_pps(...)
}
\description{
Create a new empty population play summary
}
\seealso{
Other Population Play Bot: \code{\link{pps_add_play_actions}},
  \code{\link{pps_rearrange}}
}
