% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play_bot_vs_bot}
\alias{play_bot_vs_bot}
\title{Simulate one play of the game}
\usage{
play_bot_vs_bot(game, bots, return.play.object = FALSE)
}
\arguments{
\item{game}{the game object}

\item{bots}{a list containing one bot per player}

\item{return.play.object}{By default only the outcome of the play as a one-row data frame is returned. If you set \code{return.play.object} an internal \code{play} object will be returned with more detailed information about the simulation run}
}
\description{
Simulate one play of the game
}
\seealso{
Other Bots: \code{\link{bot_eq}}, \code{\link{bot_pop}},
  \code{\link{bot_random}}, \code{\link{bot_tables}},
  \code{\link{make_bots}}, \code{\link{new_wp}}
}
