% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_pop.R
\name{pps_add_play_actions}
\alias{pps_add_play_actions}
\title{Call this function in the post.page.handler
to update the population play summary}
\usage{
pps_add_play_actions(pps, play, stage.num = play$human.stage.finished)
}
\description{
Call this function in the post.page.handler
to update the population play summary
}
\seealso{
Other Population Play Bot: \code{\link{new_pps}},
  \code{\link{pps_rearrange}}
}
