% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot_pop.R
\name{pps_rearrange}
\alias{pps_rearrange}
\title{Order pps columns naturally}
\usage{
pps_rearrange(pps)
}
\description{
If the pps is dynamically created during plays
the column order may change
}
\seealso{
Other population play functions: \code{\link{bot_pop}},
  \code{\link{new_pps}}, \code{\link{pps_add_play_actions}}
}
