% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_app.R
\name{set_wp_for_app}
\alias{set_wp_for_app}
\title{This function should be called in the appInitHandler
of your shinyEvents app}
\usage{
set_wp_for_app(wp, app = getApp(), copy = TRUE)
}
\description{
Assigns a copy of a global web play object to the
current instance of the shiny app. This means every
user has her own instance. Note that it is not possible
to have two or more web plays active in the same app.
}
\details{
The function get_wp() returns the web play object of the current app instance.
}
\seealso{
Other Web Play: \code{\link{get_wp}}, \code{\link{new_wp}},
  \code{\link{wpDevelApp}}, \code{\link{wp_developer_ui}},
  \code{\link{wp_reset}}, \code{\link{wp_set_to_play}}
}
