% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_app.R
\name{wp_developer_ui}
\alias{wp_developer_ui}
\title{A developer toolbar to your web play app}
\usage{
wp_developer_ui()
}
\description{
Returns a shiny tag list that you can add to your app$ui definition. It contains buttons to restart the experiment, edit the page rmd file in RStudio and to refresh an edited page.
Button handlers are automatically added.
}
\seealso{
Other Web Play: \code{\link{get_wp}}, \code{\link{new_wp}},
  \code{\link{set_wp_for_app}}, \code{\link{wpDevelApp}},
  \code{\link{wp_reset}}, \code{\link{wp_set_to_play}}
}
