% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{wp_reset}
\alias{wp_reset}
\title{Reset a web play to the start of a new play}
\usage{
wp_reset(wp = get_wp(), bots = wp$play$bots, human = wp$human)
}
\arguments{
\item{wp}{A web play object}

\item{bots}{You can provide new bots. By default the current bots are used again.}

\item{human}{You can define a new index of the human player. By default the current human is used.}
}
\description{
If you immediately want to start the new play. Call
\code{\link{wp_continue}} afterwards.
}
\seealso{
Other Web Play: \code{\link{get_wp}}, \code{\link{new_wp}},
  \code{\link{set_wp_for_app}}, \code{\link{wpDevelApp}},
  \code{\link{wp_copy}}, \code{\link{wp_developer_ui}},
  \code{\link{wp_set_to_play}}
}
