% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{wp_set_to_play}
\alias{wp_set_to_play}
\title{Sets the state of a web play to a play object}
\usage{
wp_set_to_play(wp, play, human = play$human)
}
\description{
Can for example be used to continue with a human
after bots played some earlier rounds
}
\seealso{
Other Web Play: \code{\link{get_wp}}, \code{\link{new_wp}},
  \code{\link{set_wp_for_app}}, \code{\link{wpDevelApp}},
  \code{\link{wp_copy}}, \code{\link{wp_developer_ui}},
  \code{\link{wp_reset}}
}
