% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp.R
\name{new_wp}
\alias{new_wp}
\title{Create a new web play object}
\usage{
new_wp(game, bots, human = 1, wpUI = "wpUI", verbose = FALSE,
  pages.dir = file.path(getwd(), "pages"), custom = list(),
  pre.page.handler = NULL, post.page.handler = NULL,
  finish.handler = wp.default.finish.handler)
}
\arguments{
\item{game}{A gtree game generated with \code{\link[gtree]{new_game}}.}

\item{bots}{A list with one bots for every player. Also add a bot for the human player. You can call \code{\link{make_bots}} to conveniently create the bots.}

\item{human}{index of the player that is played by a human in the web app.}
}
\description{
Create a new web play object
}
\seealso{
Other Bots: \code{\link{bot_eq}}, \code{\link{bot_random}},
  \code{\link{bot_tables}}, \code{\link{make_bots}},
  \code{\link{play_bot_vs_bot}}

Other Web Play: \code{\link{get_wp}},
  \code{\link{set_wp_for_app}}, \code{\link{wpDevelApp}},
  \code{\link{wp_developer_ui}}, \code{\link{wp_reset}},
  \code{\link{wp_set_to_play}}
}
